+++ cmd/urfave-cli-genflags/Makefile
@@ -1,0 +1,21 @@
+GOTEST_FLAGS ?= -v --coverprofile main.coverprofile --covermode count --cover github.com/urfave/cli/v2/cmd/urfave-cli-genflags
+GOBUILD_FLAGS ?= -x
+
+.PHONY: all
+all: test build smoke-test
+
+.PHONY: test
+test:
+	go test $(GOTEST_FLAGS) ./...
+
+.PHONY: build
+build:
+	go build $(GOBUILD_FLAGS) ./...
+
+.PHONY: smoke-test
+smoke-test: build
+	./urfave-cli-genflags --help
+
+.PHONY: show-cover
+show-cover:
+	go tool cover -func main.coverprofile
