+++ cmd/urfave-cli-genflags/generated.gotmpl
