+++ cmd/urfave-cli-genflags/generated_test.gotmpl
