+++ cmd/urfave-cli-genflags/go.mod
@@ -1,0 +1,15 @@
+module github.com/urfave/cli/v2/cmd/urfave-cli-genflags
+
+go 1.18
+
+require (
+	github.com/urfave/cli/v2 v2.11.2
+	golang.org/x/text v0.3.7
+	gopkg.in/yaml.v3 v3.0.1
+)
+
+require (
+	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
+	github.com/russross/blackfriday/v2 v2.1.0 // indirect
+	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
+)
